Minor Bug Fixes
Changed all filenames to lower case, and forced controller to view file in lower case.  This will make it so that the controllers are not case sensitive

Fixed file-link controller.  If no files are available to be downloaded, the file section will not show.  If the user has selected not to allow visitor to upload files, the upload form will not show.